feat: turn `dispose` method into `resetState` method (#7)
## Changes made
- Turned the `TimeSync.dispose` method into the `TimeSync.resetState`
method
- Updated all tests to reflect new change in functionality

## Rationale
Basically, I realized a little too late that `dispose` was too much of
an aggressive method, and would actually put some hard limitations on
what `TimeSync` is capable of doing.

As I continue to work on the React bindings, I'm realizing that I made
one really bad assumption that informed the design of both the vanilla
and React packages. That is, I assumed a consumer of `time-sync-react`
will always have access to React Context. That actually isn't true, as
Astro projects are fundamentally incompatible with context, because
Astro splits up the reactivity tree into separate islands of Reactivity.
You need a global-ish version of state in these cases, and if the global
instance is ever disposed, it fundamentally breaks the app. In test
files that would rely on the global instance, the first dispose call
would also break every other test afterwards.

I'm still working on figuring out the best way to support exposing
TimeSync as a global value (one other idea for down the line is
providing support for mixing frameworks, or at least mixing React and
Astro code). This PR is the first step towards that, but there's still a
chance that I will need to make breaking changes to support all these
use cases.