feat!: replace `refreshDate` method with `advanceTime` method (#11)
## Changes made
- Removed `refreshDate` method from `TimeSync`
- Replaced method with `advanceTime`
- Updated tests to accommodate changes

## Rationale

The more I work with TimeSync, the more I'm realizing that it's really
hard to have a good standalone library that also meets all the needs and
rules of React. Turns out, trying to bind one of the most mutable,
non-deterministic values in JavaScript to a framework that expects pure
functions is really hard

`refreshDate` (and before it, `invalidateState`) were an attempt to try
to make React happy, while exposing an API that could be used in other
contexts. The big problem, though, was that I went about it the wrong
way. There are two concerns that I wanted to handle for the React
bindings:

1. Provide a way to refresh the internal date on a fast interval so that
when a new component mounts, it's able to grab a fresh-ish value
(regardless of what intervals were used for previous components)
2. Provide a way to pipe a value from a React provider to a React hook
while maintaining a stable memory reference, so that
`useSyncExternalStore` doesn't blow up when reading it

For (1), I thought `invalidateState` would do the trick, and I added the
three notification behaviors because it felt like the right thing to do.
But it turns out, the `never` behavior is fundamentally at odds with the
entire library, and with it removed, it felt like the other two options
weren't good enough to warrant the method. So I tried turning it into
`refreshDate`, but that had its own set of wonky edge cases. Then I
tried a `setDate` that would receive an entire date as input. All of
them had their edge cases, and didn't seem great. The biggest problem,
though, was making sure that the snapshots weren't confusing – with the
`never` behavior before, the snapshot date could get wildly out of sync
with the dates that the subscribers last received

I think some sort of "invalidation-like" functionality is good, but
right now, `advanceTime` seems like the best way to do it. It has the
fewer edge cases, and seems the most obvious to outside users. Out of
all the API options, it feels like the method that's driven the least by
internal needs and is driven the most by UX

For (2), the PR briefly exposed a `lastAdvanceTarget` date property that
would always be updated when trying to advance the time, regardless of
whether the advance succeeded. That might still be necessary, but I cut
it for now. I'm going to see if the vanilla primitives are good enough
to do what the React package needs, and only pollute the API with extra
stuff if absolutely necessary