fix!: remove edge cases and add tests for TimeSync (#15)
Hoping this is the last PR before I publish v0.1.0 tonight.

## Changes made
- Revamped the function signature for `advanceTime` method and deleted
`AdvanceTimeOptions` type. Now the method only takes an optional number
- Updated `advanceTime` to reset an active interval if advancing would
dispatch to all subscribers
- Made sure that if new subscriptions are added during an update round,
that can't cause infinite loops
- Removed behavior where adding a new subscription would sometimes
refresh the internal date. Now the date is only ever updated when it's
tied to a dispatch to all subscribers (which can still be silenced for
snapshot tests)
- Updated all public and private comments where appropriate for
correctness and to avoid typos
- Updated tests to account for all code changes, and added new tests to
account for more complex interactions