fix!: remove `TimeSync.advanceTime` method (#16)
Follow-up to #11

## Changes made
- Removed the `advanceTime` method from `TimeSync` completely

## Rationale

Honestly, I think my design decisions are still being influenced by the
fact that this library started as a React library, and the vanilla
implementation only started to split off after I realized that there
were too many concerns to juggle in one place (the time management
concern and the "exposing a mutable value into a pure framework in a
render-safe way" concern).

When I first made `invalidateDate`, it was basically my attempt to
address some React-specific concerns. But, knowing that the vanilla
library could be useful when used by itself, I tried to go in the
direction of making the behavior more general. Basically, making the
most flexible version of the API I thought I needed, rather than
stepping back and deciding whether I needed it at all.

Slowly but surely, `invalidateDate` became `invalidateState`, which
became `advanceTime` with an options object, which became `advanceTime`
with a numeric value. Slowly but surely, I think I was realizing that
the method shouldn't really exist. The React use case is super niche and
can be built into the React bindings alone. The problem with
`invalidateState` and `advanceTime` is that I honestly can't tell you
what common use case they might be able to solve. They're niche
solutions that should not be baked into the core API

So, I figured that, especially with the library not being published yet,
this is the best time to remove the method entirely. It's possible that
it (or another evolution) might become necessary, but it's easier to add
a method to a published package than it is to remove it.

Plus, you know what's more simple than a four-method API? A three-method
API.