chore: add MIT license (#22)
## Changes made
- Swapped AGPL license for MIT

## Notes
- Swapped the license because the AGPL terms basically mean that any
user would have to make any derivative works open-source, which means
that companies would not touch this library with a ten-foot pole
- I used this license because it's what Coder used, and I assumed that
because they still had a way to charge for enterprise features, the
license still supported commercial use by other developers. Sounds like
that's not the case.
- Normally this should be super dicey, but because the repo has not been
published anywhere, and because it has received zero outside
contributions, this should be the one time when this should be okay