fix: remove `isSubscribed` property from context and make all other context properties readonly (#69)
## Changes made
- Deleted the `isSubscribed` property from the `SubscriptionContext`
entirely
- Redefined the `ctx.intervalLastFulfilledAt` property to be readonly
for external subscribers

## Rationale
- The more I thought about this, the potential benefits of exposing the
`isSubscribed` property didn't seem worth it, versus the confusion that
it could cause. It's the one property in the context that MUST be
mutable, but that leads to questions like "if a consumer mutates the
value, does that affect `TimeSync`?". The answer was always no, but it
was confusing. Better to remove those questions entirely
- `intervalLastFulfilledAt` is still a mutable value at runtime for
performance reasons, but we can lie to the user and tell them that it's
readonly to limit the amount of things that they try to do