feat(react): add initial support for server-side rendering (#77)
Left to do: experiment with how `useSyncExternalStore` handles
hydration, because it might actually simplify some of the new code, and
remove a bunch of edge cases

Edit: turns out, there's a lot of weird, wonky things that
`useSyncExternalStore` does to ensure data consistency. It doesn't
simplify the code I had in mind, but it might actually make some of the
existing code partially invalid. Will be addressed in a separate PR